Task/12 service layer

See merge request ussass/timetable-viewer-system!5

+++ src/test/resources/student/recreate_schema.sql
@@ -1,8 +1,8 @@
 DROP TABLE IF EXISTS students CASCADE;
 
 CREATE TABLE IF NOT EXISTS students(
-       student_id SERIAL NOT NULL PRIMARY KEY,
-       group_id INTEGER NOT NULL,
+       student_id BIGSERIAL NOT NULL PRIMARY KEY,
+       group_id BIGINT NOT NULL,
        first_name VARCHAR(255) NOT NULL,
        last_name VARCHAR(255) NOT NULL
 );
